feat: create abstract layer for deep_link_client to simplify different implementations
