+++ flutter_news_example/lib/main/main_production.dart
@@ -2,6 +2,7 @@ import 'package:ads_consent_client/ads_consent_client.dart';
 import 'package:article_repository/article_repository.dart';
 import 'package:deep_link_client/deep_link_client.dart';
 import 'package:firebase_authentication_client/firebase_authentication_client.dart';
+import 'package:firebase_deep_link_service/firebase_deep_link_service.dart';
 import 'package:firebase_notifications_client/firebase_notifications_client.dart';
 import 'package:flutter_news_example/app/app.dart';
 import 'package:flutter_news_example/main/bootstrap/bootstrap.dart';
@@ -44,7 +45,9 @@ void main() {
       );
 
       final deepLinkClient = DeepLinkClient(
-        firebaseDynamicLinks: firebaseDynamicLinks,
+        deepLinkService: FirebaseDeepLinkService(
+          firebaseDynamicLinks: firebaseDynamicLinks,
+        ),
       );
 
       final userStorage = UserStorage(storage: persistentStorage);
