+++ flutter_news_example/packages/deep_link_client/deep_link_client/lib/deep_link_client.dart
@@ -1,0 +1,2 @@
+export 'src/deep_link_client.dart';
+export 'src/deep_link_service.dart';
