+++ flutter_news_example/packages/deep_link_client/deep_link_client/lib/src/deep_link_client.dart
@@ -1,8 +1,7 @@
 import 'dart:async';
 
+import 'package:deep_link_client/deep_link_client.dart';
 import 'package:equatable/equatable.dart';
-import 'package:firebase_dynamic_links/firebase_dynamic_links.dart';
-
 import 'package:rxdart/rxdart.dart';
 
 /// {@template deep_link_client_failure}
@@ -20,20 +19,19 @@ class DeepLinkClientFailure with EquatableMixin implements Exception {
 }
 
 /// {@template deep_link_client}
-/// A client that exposes a stream of deep link URIs redirected to the app.
+/// A generic deep link client interface.
 /// {@endtemplate}
 class DeepLinkClient {
-  /// {@macro deep_link_client}
-  DeepLinkClient({FirebaseDynamicLinks? firebaseDynamicLinks})
-      : _deepLinkSubject = BehaviorSubject<Uri>() {
-    _firebaseDynamicLinks =
-        firebaseDynamicLinks ?? FirebaseDynamicLinks.instance;
-
+  /// {@macro firebase_deep_link_client}
+  DeepLinkClient({
+    required DeepLinkService deepLinkService,
+  })  : _deepLinkService = deepLinkService,
+        _deepLinkSubject = BehaviorSubject<Uri>() {
     unawaited(_getInitialLink());
-    _firebaseDynamicLinks.onLink.listen(_onAppLink).onError(_handleError);
+    _deepLinkService.deepLinkStream.listen(_onAppLink).onError(_handleError);
   }
 
-  late final FirebaseDynamicLinks _firebaseDynamicLinks;
+  final DeepLinkService _deepLinkService;
   final BehaviorSubject<Uri> _deepLinkSubject;
 
   /// Provides a stream of URIs intercepted by the app. Will emit the latest
@@ -42,7 +40,7 @@ class DeepLinkClient {
 
   Future<void> _getInitialLink() async {
     try {
-      final deepLink = await _firebaseDynamicLinks.getInitialLink();
+      final deepLink = await _deepLinkService.getInitialLink();
       if (deepLink != null) {
         _onAppLink(deepLink);
       }
@@ -51,8 +49,8 @@ class DeepLinkClient {
     }
   }
 
-  void _onAppLink(PendingDynamicLinkData dynamicLinkData) {
-    _deepLinkSubject.add(dynamicLinkData.link);
+  void _onAppLink(Uri deepLink) {
+    _deepLinkSubject.add(deepLink);
   }
 
   void _handleError(Object error, StackTrace stackTrace) {
