+++ flutter_news_example/packages/deep_link_client/deep_link_client/pubspec.yaml
@@ -1,0 +1,15 @@
+name: deep_link_client
+description: A deep link client interface
+publish_to: none
+
+environment:
+  sdk: ">=3.0.0 <4.0.0"
+
+dependencies:
+  equatable: ^2.0.3
+  mocktail: ^1.0.3
+  rxdart: ^0.27.3
+
+dev_dependencies:
+  test: ^1.21.4
+  very_good_analysis: ^5.1.0
