+++ flutter_news_example/packages/deep_link_client/firebase_deep_link_service/analysis_options.yaml
@@ -1,0 +1,1 @@
+include: package:very_good_analysis/analysis_options.5.1.0.yaml
