+++ flutter_news_example/packages/deep_link_client/firebase_deep_link_service/lib/firebase_deep_link_service.dart
@@ -1,0 +1,1 @@
+export 'src/firebase_deep_link_service.dart';
