--- flutter_news_example/packages/deep_link_client/lib/deep_link_client.dart
@@ -1,1 +1,0 @@
-export 'src/deep_link_client.dart';
